Fixed 64-bit support
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/branches/PHP_5_2@233487 c90b9560-bf6c-de11-be94-00142212c4b1

+++ main/streams/plain_wrapper.c
@@ -1301,7 +1301,7 @@ not_relative_path:
 	
 #ifdef PHP_WIN32
 	if (IS_SLASH(filename[0])) {
-		int cwd_len;
+		size_t cwd_len;
 		char *cwd;
 		cwd = virtual_getcwd_ex(&cwd_len TSRMLS_CC);
 		/* getcwd() will return always return [DRIVE_LETTER]:/) on windows. */
